Created new necessary files and added `iceCold` color.
