+++ Modulite/Extensions/UILabel+AppNameConfiguration.swift
@@ -1,0 +1,8 @@
+//
+//  UILabel+AppNameConfiguration.swift
+//  Modulite
+//
+//  Created by Gustavo Munhoz Correa on 27/08/24.
+//
+
+import Foundation
