+++ Modulite/Models/Widget/ModuleAppNameTextConfiguration.swift
@@ -1,0 +1,8 @@
+//
+//  ModuleAppNameTextConfiguration.swift
+//  Modulite
+//
+//  Created by Gustavo Munhoz Correa on 27/08/24.
+//
+
+import Foundation
