+++ Modulite/Models/Widget/WidgetStyle.swift
